SCRUM-20/SCRUM-21-SCRUM-22 refactor quiz for entity anotaions,add quizrepo with jparepo,refactor… (#18)
* refactor quiz for entity anotaions,add quizrepo with jparepo,refactored service to use new quizrepo

* add cors configuration for methods GET,POST,PUT,DELETE for front end usecase

* refactor res id to nonnull Long

This PR resolves [SCRUM-20],[SCRUM-21],[SCRUM-22] 

### Jira Tickets 
https://react-kotlin.atlassian.net/browse/SCRUM-20
https://react-kotlin.atlassian.net/browse/SCRUM-21
https://react-kotlin.atlassian.net/browse/SCRUM-22

+++ src/main/kotlin/com/reactkotlin/quiz/backend/controller/QuizController.kt
@@ -30,7 +30,7 @@ class QuizController(private val quizService: QuizService) {
 
 
     @GetMapping("/quizzes/{quizId}")
-    fun getQuiz(@PathVariable quizId: Int): ResponseEntity<QuizRes> {
+    fun getQuiz(@PathVariable quizId: Long): ResponseEntity<QuizRes> {
         return try {
             ResponseEntity.ok(quizService.getById(quizId))
         } catch (e: IllegalArgumentException) {
@@ -50,7 +50,7 @@ class QuizController(private val quizService: QuizService) {
 
 
     @PostMapping("/quizzes/{quizId}/solve")
-    fun solveQuiz(@PathVariable quizId: Int, @RequestBody quizAnswer: QuizAnswerReq): ResponseEntity<QuizAnswerRes> {
+    fun solveQuiz(@PathVariable quizId: Long, @RequestBody quizAnswer: QuizAnswerReq): ResponseEntity<QuizAnswerRes> {
 
         val answer = quizService.solve(quizId, quizAnswer.answers)
 
@@ -59,7 +59,7 @@ class QuizController(private val quizService: QuizService) {
     }
 
     @DeleteMapping("/quizzes/{quizId}")
-    fun deleteQuiz(@PathVariable quizId: Int): ResponseEntity<Unit> {
+    fun deleteQuiz(@PathVariable quizId: Long): ResponseEntity<Unit> {
 
         return try {
 
@@ -73,7 +73,7 @@ class QuizController(private val quizService: QuizService) {
 
 
     @PutMapping("/quizzes/{id}")
-    fun updateQuiz(@PathVariable id: Int,@Valid @RequestBody quiz: QuizReq): ResponseEntity<Unit> {
+    fun updateQuiz(@PathVariable id: Long,@Valid @RequestBody quiz: QuizReq): ResponseEntity<Unit> {
 
         return try {
